+++ initData/images.js
@@ -3,7 +3,7 @@ module.exports = [
     { // NOWnews墊檔圖
 
         _id: '511000000000000000000001',
-        url: 'https://legacy.nownews.com/NOWnews_default/default_terry.jpg',
+        url: 'https://legacy.nownews.com/NOWnews_default/default.png',
         CreatedBy: '530000000000000000000001',
         UpdatedBy: '530000000000000000000001',
         isDeliver: false,