+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 NOWnews-api change logs
 
 ## [Unreleased]
+### Added
+- cronjob恢復軍聞社匯入 加上圖片錯誤時存墊檔圖的機制 @appleoxxo
+- `/menus` 增加 `sort` 的搜尋條件 @esbb48
+
+### Changed
+- 新聞發布時會去更新 fb 暫存 @esbb48
 - 調整浮水印與預設圖 @esbb48
 
 ## 1.0.34 - 2017-09-28